don't track previous playing state when launching a new title
the idea was that if the user launches a game, then pauses the music using pause, then closes the game and reopens the same game, it would remember this paused state.

however, this is both confusing to the users and myself (lol). I am not sure why i added this feature. if the user wants to music to remain paused/playing, then should just set the per title config.